backend/requirements.txt
Flask==2.3.0
flask-cors==4.0.0
psycopg2-binary==2.9.6
pyserial==3.5
python-dotenv==1.0.0
